refactor: simplify OpenAI and OpenRouter settings using shared ProviderSettings component
